Exception handled

+++ 08_Customer_CRUD_with_ExceptionHandling/src/main/java/lk/ijse/spring/controller/CustomerController.java
@@ -4,6 +4,7 @@ package lk.ijse.spring.controller;
 import lk.ijse.spring.dto.CustomerDTO;
 import lk.ijse.spring.entity.Customer;
 import lk.ijse.spring.repo.CustomerRepo;
+import lk.ijse.spring.util.ResponseUtil;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,39 +29,46 @@ public class CustomerController {
 
     // get all
     @GetMapping
-    public List<CustomerDTO> getAllCustomers() {
+    public ResponseUtil getAllCustomers() {
         List<Customer> all = repo.findAll();
-        ArrayList<CustomerDTO> customerDTOS= mapper.map(all,new TypeToken<ArrayList<CustomerDTO>>(){}.getType());
-        return customerDTOS;
+        ArrayList<CustomerDTO> customerDTOS = mapper.map(all, new TypeToken<ArrayList<CustomerDTO>>() {
+        }.getType());
+        return new ResponseUtil("Ok", "Successfully Loaded", customerDTOS);
     }
 
     // find
     @GetMapping(params = {"id"})
-    public CustomerDTO findCustomer(String id) {
-        Customer c = repo.findById(id).get();
-        CustomerDTO map = mapper.map(c, CustomerDTO.class);
-        return map;
+    public ResponseUtil findCustomer(String id) {
+        if (!repo.existsById(id)) throw new RuntimeException("Id not exists !");
+
+        CustomerDTO map = mapper.map(repo.findById(id).get(), CustomerDTO.class);
+        return new ResponseUtil("Ok", "Successfully Searched", map);
     }
 
     // add
     @PostMapping
-    public void saveCustomer(@ModelAttribute CustomerDTO dto) {
-        Customer map = mapper.map(dto, Customer.class);
-        repo.save(map);
+    public ResponseUtil saveCustomer(@ModelAttribute CustomerDTO dto) {
+        if (repo.existsById(dto.getId())) throw new RuntimeException("Error, Already added!");
+
+        repo.save(mapper.map(dto, Customer.class));
+        return new ResponseUtil("Ok", "Successfully Added", dto);
     }
 
     // update
     @PutMapping
-    public void updateCustomer(@RequestBody CustomerDTO dto) {
-        if (repo.findById(dto.getId()).isPresent()) { // check if customer exists
-            Customer map = mapper.map(dto, Customer.class);
-            repo.save(map);
-        }
+    public ResponseUtil updateCustomer(@RequestBody CustomerDTO dto) {
+        if (!repo.existsById(dto.getId())) throw new RuntimeException("Id not exists !");
+
+        repo.save(mapper.map(dto, Customer.class));
+        return new ResponseUtil("Ok", "Successfully Updated", dto);
     }
 
     // delete
     @DeleteMapping(params = {"id"})
-    public void deleteCustomer(String id) {
+    public ResponseUtil deleteCustomer(String id) {
+        if (!repo.existsById(id)) throw new RuntimeException("Id not exists !");
+
         repo.deleteById(id);
+        return new ResponseUtil("Ok", "Successfully Deleted", id);
     }
 }
